Make library crate so that controller tests work better
