+++ mini-projects/01-device-driver/laptop-interface/src/lib_serial_ffi.rs
@@ -1,4 +1,4 @@
-#![allow(non_camel_case_types)]
+#![allow(non_camel_case_types, dead_code)]
 use log::{debug, error, info, trace, warn};
 use std::ffi::{CStr, CString};
 use std::os::raw::{c_char, c_int, c_uint, c_void};
